Revert "Corrections for using water quality modules with group macros."
This reverts commit 704baf7b64e9399fdac6c97bbae13b965c9556a9.

+++ crhmcode/src/modules/waterquality/ClassGrow_crops_annually.cpp
@@ -337,7 +337,7 @@ void ClassGrow_crops_annually::run(void) {
                 }
                 if (ObsCnt_fertperiod >= hh) { // file open
                     declputparam("*", "fertperiod", "(d)", &fertperiod);
-                    const_cast<double*> (fertperiod)[hh] = 0; // set by this module
+                    const_cast<long*> (fertperiod)[hh] = 0; // set by this module
                 }
                 //        if(ObsCnt_litterperiod >= hh){ // file open  check ????
                 declputparam("*", "litterperiod", "(d)", &litterperiod);
